Push responsibility for knowing if a func is implemented to backend.
Rather than having the top-level BLEDevice know which backend implements
what, require that the backend implement the same interface - the
BLEBackend. By default, all of the functions in the interface will raise
a NotImplementedError. If a backend is capable of that feature, it just
overrides the function with its actual implementation. If it doesn't
provide one (e.g. gatttool doesn't implement get_rssi), it will raise an
exception at runtime.