Store characteristics on GATTTool device so they aren't requested twice.
Due to one-device-at-a-time limitation on the GATTTool backend, the
query for characteristics actually happens in the backend class. The
list was being stored on the backend but not shared with the device
instance, so if you called any other function that needed to find the
handle for a UUID (e.g. subscribing or reading or writing), it would
query for the characteristics again.

This makes sure the characteristic are also copied to the device
instance so the device can read from the cache.