Connect to non-advertising device support
Previously, if we tried to connect to a device that had stopped
advertising then we couldn’t connect to any device if it were
advertising or not. (Ex: You properly issued a scan and found the
address you cared about but, by the time you were able to issue a
connect, the device had stopped advertising. After this operation you
are unable to do any other connections.)

This change makes it so that we follow the bluegiga documentation that
suggests that we use the End Procedure command to cancel the connection
process. The current code “times out” but doesn’t communicate that
intention to the bluegiga system which leaves it hanging.