Resetting backend cleanly
Instead of trying to nicely shut down all of the individual things on
the bluegiga system we just request do a clean restart. It may seem
heavy handed but that is really the best way to ensure that our state
machine and the one on the embedded system are in sync.

I was driven to this after solution after having a number of issues
when I was trying to iterate quickly on some scripts. I found myself
having to physically remove the device fairly frequently due to the
fact that some exception would put the chip in a state that the code
couldn’t properly recover it from. That was unacceptable and I now no
longer have to do that.