Attempt to gracefully stop lescan to leave adapter in a happy state.
If `hcitool lescan` is not allowed a graceful exit, it can leave the
adapter (or BlueZ) stack in a state where it reports a failure to set
scan parameters, and you have to reset the adapter. This change sends a
SIGINT to the process and waits for it to terminate before continuing,
allowing multiple scans without resetting.

If the run_as_root option was True, however, we can't send SIGINT as a
normal user.

Fixes #88