Wait half a second after restarting the USB adapter.
This can be necessary to avoid the _open_serial_port call to fail. Else
it will raise an exception in the first iteration and abort with a
NotConnectedError.

Note that 250ms seems to be sufficient on some setups [0], but at least
on my Windows setup on a Virtual Machine the half second is necessary.

[0] https://github.com/peplin/pygatt/issues/118#issuecomment-292058009