Do not abort on first attempt in _open_serial_port.
During bgapi.start the BLE adapter is first reset before connecting to
it again and starting a receiver thread. Sometimes though the adapter is
not ready yet on the first connection attempt, and breaking out early
results in an unnecessarily raised exception.

Allowing _open_serial_port to retry makes it robust. Even in the worst
case where there actually is no BLE adapter connected the 10 retry
attempts only take a moment, so taking this moment longer is not too
bad.