Add debug log print to log failed connection attempts.
Before _open_serial_port would try 10 attempts to establish a
connection, and if that failed raise an exception stating 'Unable to
reconnect with USB device after rebooting'.
This does not actually communicate what exactly went wrong though. To
alleviate this we simply add a debug print with the actual exception
that causes the connection establishments to fail.