Clarify USB (re)connection debug prints.
_open_serial_port prints that it attempts to connect or failed to
connect _after_ the USB device was restarted. This is somewhat
misleading as the start method actually calls _open_serial_port twice,
and the first time before the USB device is restarted. To make it more
clear we simply do not refer to the device restart in _open_serial_port,
and instead simply announce the restart with an additional debug print in
the start method.