Remove default of 200 search depth of pexpect buffer
Should someone have the need to only look for subset of messages they can set
limit on per application basis. `200` is not an appropriate default as
`max_read` is set to `2000` which means we can discard 90% of the messages we
receive -- not a good default behavior.

This commit also adds option to tune the max_read. This should be tuned in
unison with `search_window_size`.