char_read and char_write were non-functional
I had to re-write them to use managed objects. But then, based on the
way managed objects are queried, things were REAL slow. So we cache the
managed objects for ones we have already searched for.  There is
probably a lower layer to cache things that might work better but this
solved all the issues that I had.