Final Process class bugfixes
I was able to get this whole system only to find out that even if the
whole process was dedicated to consuming bluetooth packets at the rate
I was consuming them and on the platform that I am on it wasn’t enough.
I believe the glib garbage collection wasn’t being given enough time to
run but regardless it didn’t solve the problem. I am leaving these
process classes around because they are an interesting way to offload
processing load and might be useful in the future.