Fix loop over a mutable dict during BGAPI shutdown
This fixes a RuntimeError reported in #306.

There's a circular dependency between the backend and device classes.
When stopping the BGAPI backend, we call `disconnect` on each device.
When a disconnection event is received from the BLED112 adapter, the
device is removed from `self._connections` - the same dict we are
iterating over.